Refactor billing data processing: consolidate asset imports and create new asset files for aggregates, ingestion, and reports
