+++ pipeline/assets/__init__.py
@@ -1,0 +1,18 @@
+from .aggregates import (
+    daily_aggregates,
+    region_aggregates,
+    service_aggregates,
+    user_aggregates,
+)
+from .ingest import billing_db, billing_files
+from .reports import billing_insights
+
+__all__ = [
+    "billing_files",
+    "billing_db",
+    "daily_aggregates",
+    "region_aggregates",
+    "service_aggregates",
+    "user_aggregates",
+    "billing_insights",
+]
