+++ pipeline/assets/ingest.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from dagster import AssetExecutionContext, MetadataValue, asset
 from dagster_duckdb import DuckDBResource
 
-from .utils import get_table_preview
+from ..utils import get_table_preview
 
 
 @asset
